build: minor ws changes
