+++ basic-package/{{cookiecutter.github_repo}}/{{cookiecutter._vscode_ws_name}}.code-workspace
@@ -359,6 +359,22 @@
 				]
 			},
 			{
+				"label": "mypy: generate file stub",
+				"type": "shell",
+				"command": "conda",
+				"args": [
+					"run",
+					"--no-capture-output",
+					"-n",
+					"{{cookiecutter.__pkg_import_name}}",
+					"stubgen",
+					"${fileDirname}/${fileBasenameNoExtension}.py",
+					"-o",
+					"typings",
+					"--include-private"
+				]
+			},
+			{
 				"label": "cookiecutter: replay",
 				"type": "shell",
 				"command": "scripts/replay/replay.sh",
