+++ datajoint-workflow/{{cookiecutter.github_repo}}/pyproject.toml
@@ -4,6 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "{{cookiecutter.__pkg_name}}"
+dynamic = ["version"]
 description = "A {{cookiecutter.organization}} Workflow for {{cookiecutter.workflow}}"
 readme = "README.md"
 requires-python = ">={{cookiecutter.python_version}}"
@@ -21,7 +22,6 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
 ]
-dynamic = ["version"]
 dependencies = [
     "cryptography",                                                                       # datajoint
     "ipython",                                                                            # datajoint
@@ -53,7 +53,16 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-dev = ["black", "isort", "flake8", "pre-commit", "nox", "mypy", "notebook"]
+dev = [
+    "black",
+    "isort",
+    "flake8",
+    "pre-commit",
+    "nox",
+    "mypy",
+    "notebook",
+    "ipykernel",
+]
 doc = [
     "mkdocs",
     "mkdocstrings-python",
@@ -66,7 +75,10 @@ doc = [
     "mike",
 ]
 test = ["pytest", "faker"]
-sciops = ["ipykernel", "djsciops"]
+sciops = [
+    "dj-wft-cim @ git+https://github.com/{{cookiecutter.github_user}}/wt-causality-in-motion",
+    "djsciops",
+]
 
 [project.scripts]
 {{cookiecutter.__pkg_import_name}}_entrypoint = "{{cookiecutter.__pkg_import_name}}.populate.entrypoint:cli"
@@ -81,9 +93,15 @@ DataJoint = "https://docs.datajoint.org/"
 [tool.setuptools.dynamic]
 version = { attr = "{{cookiecutter.__pkg_import_name}}.__version__" }
 
+[tool.setuptools]
+package-data = { "*" = ["*.pyi", "py.typed"] }
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
 [tool.pytest.ini_options]
 minversion = "6.0"
-addopts = "-ra -q"
+addopts = "-rA"
 testpaths = ["tests"]
 
 [tool.black]
@@ -116,7 +134,9 @@ exclude = [
     "**/.ipynb_checkpoints",
     "**/__pycache__",
     "src/experimental",
+    "src/scratch",
     "notebooks",
+    "tests"
 ]
 ignore = ["**/*.ipynb", "notebooks"]
 stubPath = "typings"
